+++ src/main/java/edu/metrostate/booknow/Utils/RestaurantUIManager.java
@@ -15,6 +15,7 @@ import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
+import javafx.util.Callback;
 
 import java.awt.*;
 import java.io.File;
@@ -48,7 +49,6 @@ public class RestaurantUIManager {
     private static final String TEXT_FILL_GRAY = "-fx-text-fill: gray;";
     private static final String TEXT_FILL_WHITE = "-fx-text-fill: white;";
     private static final String TEXT_FILL_GREEN = "-fx-text-fill: green;";
-    private static final String TEXT_FILL_ORANGE = "-fx-text-fill: orange;";
     private static final String TEXT_FILL_RED = "-fx-text-fill: red;";
     private static final String BACKGROUND_COLOR_BLUE = "-fx-background-color: #003580;";
 
@@ -60,7 +60,7 @@ public class RestaurantUIManager {
     private static final String RESTAURANT_BUTTONS_STYLE = SPACING_STYLE_5;
 
     private static final String REVIEWS_AND_RATING_STYLE = SPACING_STYLE_5;
-    private static final String REVIEW_BUTTON_STYLE = FONT_WEIGHT_BOLD + TEXT_FILL_BLACK;
+    private static final String REVIEW_BUTTON_STYLE = FONT_WEIGHT_BOLD + TEXT_FILL_WHITE + BACKGROUND_COLOR_BLUE;
 
     private static final String RATING_STYLE = "-fx-padding: 5;" + BACKGROUND_COLOR_BLUE + FONT_WEIGHT_BOLD + FONT_SIZE_14 + TEXT_FILL_WHITE;
     private static final String BUTTON_STYLE = FONT_WEIGHT_BOLD + TEXT_FILL_BLACK;
@@ -72,25 +72,10 @@ public class RestaurantUIManager {
 
     private static final String TABLE_VIEW_STYLE = "-fx-border-color: lightgray; -fx-padding: 10;";
 
-    private static final String REVIEW_BOX_STYLE = "-fx-padding: 10; -fx-spacing: 5; -fx-border-color: lightgray; -fx-border-width: 1;";
-    private static final String USERNAME_LABEL_STYLE = FONT_SIZE_14 + TEXT_FILL_BLACK;
-    private static final String RATING_LABEL_STYLE = FONT_SIZE_14 + TEXT_FILL_ORANGE;
-    private static final String FEEDBACK_LABEL_STYLE = FONT_SIZE_12 + TEXT_FILL_BLACK;
-    private static final String DATE_LABEL_STYLE = FONT_SIZE_12 + TEXT_FILL_GRAY;
+    private static final String REVIEW_BOX_STYLE = "-fx-padding: 15; -fx-spacing: 8; -fx-border-color: #d3d3d3; -fx-border-width: 1; -fx-border-radius: 8; -fx-background-color: #f9f9f9; -fx-background-radius: 8; -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.1), 5, 0, 3, 3);";
 
-
-    /**
-     * Populates a given VBox with a list of restaurants.
-     *
-     * @param restaurantListVBox The VBox container to be populated with restaurant entries.
-     * @param restaurants The list of restaurants to display.
-     * @param onReadReviews A Consumer function to handle the action of reading reviews for a restaurant.
-     * @param onViewMenu A Consumer function to handle the action of viewing the menu of a restaurant.
-     * @param onShowAvailability A Consumer function to handle the action of showing availability for a restaurant.
-     */
     public void populateRestaurantListVBox(VBox restaurantListVBox, List<Restaurant> restaurants, Consumer<Restaurant> onReadReviews, Consumer<Restaurant> onViewMenu, Consumer<Restaurant> onShowAvailability) {
         restaurantListVBox.getChildren().clear();
-
         for (Restaurant restaurant : restaurants) {
             HBox restaurantBox = createRestaurantBox(restaurant, onReadReviews, onViewMenu, onShowAvailability);
             restaurantListVBox.getChildren().add(restaurantBox);
@@ -99,15 +84,11 @@ public class RestaurantUIManager {
 
     private HBox createRestaurantBox(Restaurant restaurant, Consumer<Restaurant> onReadReviews, Consumer<Restaurant> onViewMenu, Consumer<Restaurant> onShowAvailability) {
         HBox restaurantBox = new HBox();
-        // spacing + border padding
         restaurantBox.setStyle(RESTAURANT_BOX_STYLE);
-
         ImageView restaurantImage = createRestaurantImageView(restaurant.getImagePath());
         VBox restaurantDetails = createRestaurantDetails(restaurant);
         VBox restaurantButtons = createRestaurantButtons(restaurant, onReadReviews, onViewMenu, onShowAvailability);
-
         restaurantBox.getChildren().addAll(restaurantImage, restaurantDetails, restaurantButtons);
-
         return restaurantBox;
     }
 
@@ -115,64 +96,51 @@ public class RestaurantUIManager {
         ImageView restaurantImage = new ImageView();
         restaurantImage.setFitHeight(150);
         restaurantImage.setFitWidth(200);
-
         String imageFilePath = "/" + imagePath;
         try {
             restaurantImage.setImage(new Image(getClass().getResource(imageFilePath).toExternalForm()));
         } catch (NullPointerException e) {
             restaurantImage.setImage(new Image(getClass().getResource("/images/default.jpg").toExternalForm()));
         }
-
         return restaurantImage;
     }
 
     private VBox createRestaurantDetails(Restaurant restaurant) {
         VBox restaurantDetails = new VBox();
         restaurantDetails.setStyle(RESTAURANT_DETAILS_STYLE);
-
         Label nameLabel = new Label(restaurant.getName());
         nameLabel.setStyle(RESTAURANT_NAME_STYLE);
-
         Label locationLabel = new Label(restaurant.getCity());
         locationLabel.setStyle(RESTAURANT_LOCATION_STYLE);
-
         Label descriptionLabel = new Label(restaurant.getDescription());
-
         restaurantDetails.getChildren().addAll(nameLabel, locationLabel, descriptionLabel);
         HBox.setHgrow(restaurantDetails, Priority.ALWAYS);
-
         return restaurantDetails;
     }
 
     private VBox createRestaurantButtons(Restaurant restaurant, Consumer<Restaurant> onReadReviews, Consumer<Restaurant> onViewMenu, Consumer<Restaurant> onShowAvailability) {
         VBox restaurantButtons = new VBox();
         restaurantButtons.setStyle(RESTAURANT_BUTTONS_STYLE);
-
         HBox reviewsAndRating = createReviewsAndRatingBox(restaurant, onReadReviews);
         restaurantButtons.getChildren().add(reviewsAndRating);
-
         Button menuButton = createButton("View Menu", onViewMenu, restaurant);
+        menuButton.setStyle(REVIEW_BUTTON_STYLE); // Apply the same style
         Button availabilityButton = createButton("Show Availability", onShowAvailability, restaurant);
-
+        availabilityButton.setStyle(REVIEW_BUTTON_STYLE); // Apply the same style
         restaurantButtons.getChildren().addAll(menuButton, availabilityButton);
-
         return restaurantButtons;
     }
 
     private HBox createReviewsAndRatingBox(Restaurant restaurant, Consumer<Restaurant> onReadReviews) {
         HBox reviewsAndRating = new HBox();
         reviewsAndRating.setStyle(REVIEWS_AND_RATING_STYLE);
-
         Button reviewsButton = new Button("Read Reviews");
         reviewsButton.setStyle(REVIEW_BUTTON_STYLE);
         reviewsButton.setOnAction(event -> onReadReviews.accept(restaurant));
-
         String averageRating = reviewService.getAverageRating(restaurant.getRestaurantId());
         Label ratingLabel = new Label(averageRating);
         ratingLabel.setStyle(RATING_STYLE);
-
         reviewsAndRating.getChildren().addAll(reviewsButton, ratingLabel);
-
         return reviewsAndRating;
     }
 
@@ -183,7 +151,6 @@ public class RestaurantUIManager {
         return button;
     }
 
-    @SuppressWarnings("unchecked")
     public TableView<Table> createTableView(Restaurant restaurant, LocalDate selectedDate, int selectedGuests, Consumer<Table> onReserveTable, Consumer<String> onTimeSlotSelected) {
         TableView<Table> tableView = new TableView<>();
 
@@ -192,22 +159,37 @@ public class RestaurantUIManager {
         TableColumn<Table, Double> priceColumn = new TableColumn<>("Price");
         TableColumn<Table, String> timeSlotColumn = new TableColumn<>("Available Time Slots");
         TableColumn<Table, Void> reserveColumn = new TableColumn<>("Reserve");
-
         tableNumberColumn.setCellValueFactory(new PropertyValueFactory<>("tableNumber"));
         numberOfSeatsColumn.setCellValueFactory(new PropertyValueFactory<>("numberOfSeats"));
         priceColumn.setCellValueFactory(new PropertyValueFactory<>("bookingFee"));
 
-        priceColumn.setCellFactory(column -> new TableCell<Table, Double>() {
+        priceColumn.setCellFactory(createPriceCellFactory());
+        reserveColumn.setCellFactory(createReserveCellFactory(onReserveTable));
+        numberOfSeatsColumn.setCellFactory(createSeatsCellFactory());
+        timeSlotColumn.setCellFactory(createTimeSlotCellFactory(restaurant, selectedDate, selectedGuests, onTimeSlotSelected));
+
+        Collections.addAll(tableView.getColumns(), tableNumberColumn, numberOfSeatsColumn, priceColumn, timeSlotColumn, reserveColumn);
+
+        tableView.setStyle(TABLE_VIEW_STYLE);
+        tableView.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY_FLEX_LAST_COLUMN);
+        List<Table> availableTables = tableService.getAvailableTables(restaurant.getRestaurantId(), selectedDate, selectedGuests);
+        tableView.setItems(FXCollections.observableArrayList(availableTables));
+
+        return tableView;
+    }
+
+    private Callback<TableColumn<Table, Double>, TableCell<Table, Double>> createPriceCellFactory() {
+        return column -> new TableCell<Table, Double>() {
+            private static final String BOOKING_POLICY_MESSAGE = "Note: Booking fee is non-refundable if canceled within 24 hours.";
             private final VBox vbox = new VBox();
             private final Label priceLabel = new Label();
-            private final Label bookingPolicyLabel = new Label("Note: Booking fee is non-refundable if canceled within 24 hours.");
-
+            // Just for looks actually does nothing
+            private final Label bookingPolicyLabel = new Label(BOOKING_POLICY_MESSAGE);
             {
                 bookingPolicyLabel.setStyle(BOOKING_POLICY_STYLE);
                 vbox.getChildren().addAll(priceLabel, bookingPolicyLabel);
                 vbox.setSpacing(2);
             }
-
             @Override
             protected void updateItem(Double item, boolean empty) {
                 super.updateItem(item, empty);
@@ -219,11 +201,12 @@ public class RestaurantUIManager {
                     setGraphic(vbox);
                 }
             }
-        });
+        };
+    }
 
-        reserveColumn.setCellFactory(col -> new TableCell<Table, Void>() {
+    private Callback<TableColumn<Table, Void>, TableCell<Table, Void>> createReserveCellFactory(Consumer<Table> onReserveTable) {
+        return col -> new TableCell<Table, Void>() {
             private final Button reserveButton = new Button("Reserve");
-
             {
                 reserveButton.setStyle(RESERVE_BUTTON_STYLE);
                 reserveButton.setOnAction(event -> {
@@ -231,7 +214,6 @@ public class RestaurantUIManager {
                     onReserveTable.accept(table);
                 });
             }
-
             @Override
             protected void updateItem(Void item, boolean empty) {
                 super.updateItem(item, empty);
@@ -241,9 +223,11 @@ public class RestaurantUIManager {
                     setGraphic(reserveButton);
                 }
             }
-        });
+        };
+    }
 
-        numberOfSeatsColumn.setCellFactory(col -> new TableCell<Table, Integer>() {
+    private Callback<TableColumn<Table, Integer>, TableCell<Table, Integer>> createSeatsCellFactory() {
+        return col -> new TableCell<Table, Integer>() {
             @Override
             protected void updateItem(Integer item, boolean empty) {
                 super.updateItem(item, empty);
@@ -262,11 +246,12 @@ public class RestaurantUIManager {
                     setGraphic(guestIcons);
                 }
             }
-        });
+        };
+    }
 
-        timeSlotColumn.setCellFactory(column -> new TableCell<Table, String>() {
+    private Callback<TableColumn<Table, String>, TableCell<Table, String>> createTimeSlotCellFactory(Restaurant restaurant, LocalDate selectedDate, int selectedGuests, Consumer<String> onTimeSlotSelected) {
+        return column -> new TableCell<Table, String>() {
             private final ComboBox<String> timeSlotComboBox = new ComboBox<>();
-
             @Override
             protected void updateItem(String item, boolean empty) {
                 super.updateItem(item, empty);
@@ -275,27 +260,16 @@ public class RestaurantUIManager {
                 } else {
                     Table table = getTableView().getItems().get(getIndex());
                     List<String> availableSlots = tableService.getAvailableTimeSlots(restaurant, selectedDate, selectedGuests, table);
-
                     timeSlotComboBox.setItems(FXCollections.observableArrayList(availableSlots));
                     timeSlotComboBox.setOnAction(event -> {
                         String selectedSlot = timeSlotComboBox.getSelectionModel().getSelectedItem();
                         onTimeSlotSelected.accept(selectedSlot);
                     });
                     timeSlotComboBox.getSelectionModel().selectFirst();
-
                     setGraphic(timeSlotComboBox);
                 }
             }
-        });
-
-        Collections.addAll(tableView.getColumns(), tableNumberColumn, numberOfSeatsColumn, priceColumn, timeSlotColumn, reserveColumn);
-        tableView.setStyle(TABLE_VIEW_STYLE);
-        tableView.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY_FLEX_LAST_COLUMN);
-
-        List<Table> availableTables = tableService.getAvailableTables(restaurant.getRestaurantId(), selectedDate, selectedGuests);
-        tableView.setItems(FXCollections.observableArrayList(availableTables));
-
-        return tableView;
+        };
     }
 
     public void displayReviews(VBox reviewsOverlay, List<Review> reviews) {
@@ -319,24 +293,48 @@ public class RestaurantUIManager {
     }
 
     private VBox createReviewBox(Review review) {
-        Label usernameLabel = new Label(review.getUsername().substring(0, 2) + "****");
-        usernameLabel.setStyle(USERNAME_LABEL_STYLE);
+        // Main container for the review
+        VBox reviewBox = new VBox();
+        // #f9f9f9 very light shade of grey #d3d3d3 light grey color
+        reviewBox.setStyle("-fx-padding: 20; -fx-spacing: 8; -fx-border-color: #d3d3d3; -fx-border-width: 1; -fx-border-radius: 8; -fx-background-color: #f9f9f9; -fx-background-radius: 8;");
 
+        // Load the close icon image
+        ImageView closeIcon = new ImageView(new Image(getClass().getResource("/images/close_icon.png").toExternalForm()));
+        closeIcon.setFitWidth(12);  // Adjust icon size
+        closeIcon.setFitHeight(12);
+
+        // Create a close button and add the close icon to it
+        Button closeButton = new Button();
+        closeButton.setGraphic(closeIcon);
+        closeButton.setStyle("-fx-background-color: transparent; -fx-cursor: hand; -fx-padding: 2;");
+
+        closeButton.setOnAction(event -> {
+            // Remove this review box from the reviews overlay
+            ((VBox) reviewBox.getParent()).getChildren().remove(reviewBox);
+        });
+
+        // Add close button to the top-right corner
+        HBox topBar = new HBox(closeButton);
+        topBar.setStyle("-fx-alignment: top-right;"); // Align button to the top-right corner
+
+        // Labels for the review details
+        Label usernameLabel = new Label(review.getUsername().substring(0, 2) + "****");
+        usernameLabel.setStyle("-fx-font-size: 14px; -fx-text-fill: black;");
+
+        // To hide full username of someone else
         String stars = "★".repeat(review.getRating()) + "☆".repeat(5 - review.getRating());
         Label ratingLabel = new Label(stars);
-        ratingLabel.setStyle(RATING_LABEL_STYLE);
+        ratingLabel.setStyle("-fx-font-size: 14px; -fx-text-fill: orange;");
 
         Label feedbackLabel = new Label(review.getFeedback());
-        feedbackLabel.setStyle(FEEDBACK_LABEL_STYLE);
+        feedbackLabel.setStyle("-fx-font-size: 12px; -fx-text-fill: black;");
 
         Label dateLabel = new Label("Date of Experience: " + review.getDateOfExperience());
-        dateLabel.setStyle(DATE_LABEL_STYLE);
+        dateLabel.setStyle("-fx-font-size: 12px; -fx-text-fill: gray;");
 
-        VBox reviewBox = new VBox(usernameLabel, ratingLabel, feedbackLabel, dateLabel);
-        reviewBox.setStyle(REVIEW_BOX_STYLE);
+        reviewBox.getChildren().addAll(topBar, usernameLabel, ratingLabel, feedbackLabel, dateLabel);
         return reviewBox;
     }
-
 
     public void viewMenu(Restaurant restaurant) {
         try {
@@ -377,6 +375,4 @@ public class RestaurantUIManager {
     private InputStream loadPDF(String menuName) {
         return getClass().getResourceAsStream("/menus/" + menuName);
     }
-
-
 }